requirements.txt
livekit-agents>=1.0.0
livekit-plugins-soniox>=0.3.0
livekit-plugins-groq>=0.3.0
livekit-plugins-elevenlabs>=0.8.0
livekit-plugins-silero>=0.7.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
structlog>=24.1.0
